+++ client/synapseJavaClient/src/main/java/org/sagebionetworks/client/SynapseClientImpl.java
@@ -4730,7 +4730,7 @@ public class SynapseClientImpl extends BaseClientImpl implements SynapseClient {
 	public Jwt<JwsHeader,Claims> getUserInfoAsJSONWebToken() throws SynapseException {
 		Map<String,String> requestHeaders = new HashMap<String,String>();
 		requestHeaders.put(AuthorizationConstants.AUTHORIZATION_HEADER_NAME, getAuthorizationHeader());
-		SimpleHttpResponse response = signAndDispatchSynapseRequest(
+		SimpleHttpResponse response = dispatchSynapseRequest(
 				getAuthEndpoint(), AUTH_OAUTH_2_USER_INFO, GET, null, requestHeaders, null);
 		if (!ClientUtils.is200sStatusCode(response.getStatusCode())) {
 			ClientUtils.throwException(response.getStatusCode(), response.getContent());
@@ -4787,14 +4787,17 @@ public class SynapseClientImpl extends BaseClientImpl implements SynapseClient {
 	}
 
 	@Override
-	public void revokeTokenURLEncoded(String token) throws SynapseException {
+	public void revokeTokenURLEncoded(String token) throws SynapseException, UnsupportedEncodingException {
 		ValidateArgument.required(token, "token");
 		Map<String, String> headers = new HashMap<String, String>();
+		// Note, the authorization header is expected to be the Basic Authorization
+		// credentials (client id & secret) for the OAuth client making this request
 		headers.put(AuthorizationConstants.AUTHORIZATION_HEADER_NAME, getAuthorizationHeader());
-		headers.put(CONTENT_TYPE, "application/x-www-form-urlencoded");
-		String requestBody="token="+token;
+		String charset="UTF-8";
+		headers.put(CONTENT_TYPE, "application/x-www-form-urlencoded; charset="+charset);
+		String requestBody="token="+URLEncoder.encode(token, charset);
 		
-		SimpleHttpResponse response = signAndDispatchSynapseRequest(getAuthEndpoint(),
+		SimpleHttpResponse response = dispatchSynapseRequest(getAuthEndpoint(),
 				AUTH_OAUTH_2 + REVOKE, POST, requestBody, headers, null);
 		ClientUtils.checkStatusCodeAndThrowException(response);
 	}
