+++ services/repository-managers/src/test/java/org/sagebionetworks/repo/web/controller/JSONEntityHttpMessageConverterTest.java
@@ -5,15 +5,11 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static  org.sagebionetworks.repo.web.controller.JSONEntityHttpMessageConverter.convertFormEncodedDataToJSONString;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.io.StringReader;
 import java.nio.charset.Charset;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -24,16 +20,10 @@ import org.junit.Test;
 import org.mockito.Mockito;
 import org.sagebionetworks.repo.model.Entity;
 import org.sagebionetworks.repo.model.ErrorResponse;
-import org.sagebionetworks.repo.model.ExampleEntity;
 import org.sagebionetworks.repo.model.Project;
 import org.sagebionetworks.repo.model.schema.CreateSchemaRequest;
-import org.sagebionetworks.repo.model.schema.JsonSchema;
-import org.sagebionetworks.schema.adapter.JSONArrayAdapter;
 import org.sagebionetworks.schema.adapter.JSONEntity;
 import org.sagebionetworks.schema.adapter.JSONObjectAdapter;
-import org.sagebionetworks.schema.adapter.JSONObjectAdapterException;
-import org.sagebionetworks.schema.adapter.org.json.EntityFactory;
-import org.sagebionetworks.schema.adapter.org.json.JSONArrayAdapterImpl;
 import org.sagebionetworks.schema.adapter.org.json.JSONObjectAdapterImpl;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpInputMessage;
@@ -45,8 +35,6 @@ import com.amazonaws.util.StringInputStream;
 
 
 public class JSONEntityHttpMessageConverterTest {
-
-	
 	Project project;
 	
 	HttpOutputMessage mockOutMessage;
@@ -54,9 +42,7 @@ public class JSONEntityHttpMessageConverterTest {
 	ByteArrayOutputStream outStream;
 	HttpHeaders mockHeaders;
 	JSONEntityHttpMessageConverter converter;
-	
 
-	
 	@Before
 	public void before() throws IOException{	
 		project = new Project();
@@ -105,23 +91,6 @@ public class JSONEntityHttpMessageConverterTest {
 	}
 	
 	@Test
-	public void testConvertFormEncodedDataToJSONString_HappyCase() {
-		assertEquals("{\"foo\":\"bar\",\"bar\":\"baz\"}", convertFormEncodedDataToJSONString("foo=bar&bar=baz"));
-	}
-	
-	@Test
-	public void testConvertFormEncodedDataToJSONString_EmptyString() {
-		assertEquals("{}", convertFormEncodedDataToJSONString(""));
-	}
-	
-	@Test
-	public void testConvertFormEncodedDataToJSONString_InvalidInput() {
-		assertThrows(IllegalArgumentException.class, () -> {
-			convertFormEncodedDataToJSONString("garbage");
-		});
-	}
-	
-	@Test
 	public void testRoundTripWithPlainTextMediaType() throws HttpMessageNotWritableException, IOException{
 		// Write it out.
 		converter.write(project, MediaType.TEXT_PLAIN, mockOutMessage);
@@ -166,68 +135,6 @@ public class JSONEntityHttpMessageConverterTest {
 		error.setReason("foo");
 		assertEquals("foo", JSONEntityHttpMessageConverter.convertEntityToPlainText(error));
 	}
-
-	@Test 
-	public void testReadToString() throws IOException{
-		String value = "This string should make a round trip!";
-		StringReader reader = new StringReader(value);
-		String clone = JSONEntityHttpMessageConverter.readToString(reader);
-		assertEquals(value, clone);
-	}
-	
-	@Test
-	public void testReadEntity() throws JSONObjectAdapterException, IOException{
-		ExampleEntity entity = new ExampleEntity();
-		entity.setName("name");
-		// this version requires a class name fo the entity type.
-		entity.setDoubleList(new ArrayList<Double>());
-		entity.getDoubleList().add(123.45);
-		entity.getDoubleList().add(4.56);
-		// To string
-		String jsonString =EntityFactory.createJSONStringForEntity(entity);
-		StringReader reader = new StringReader(jsonString);
-		ExampleEntity clone = (ExampleEntity) JSONEntityHttpMessageConverter.readEntity(reader);
-		assertEquals(entity, clone);
-	}
-	
-	@Test (expected=JSONObjectAdapterException.class)
-	public void testReadEntityNullType() throws JSONObjectAdapterException, IOException{
-		ExampleEntity entity = new ExampleEntity();
-		entity.setName("name");
-		// this version requires a class name fo the entity type.
-		entity.setConcreteType(null);
-		entity.setDoubleList(new ArrayList<Double>());
-		entity.getDoubleList().add(123.45);
-		entity.getDoubleList().add(4.56);
-		// To string
-		String jsonString =EntityFactory.createJSONStringForEntity(entity);
-		StringReader reader = new StringReader(jsonString);
-		ExampleEntity clone = (ExampleEntity) JSONEntityHttpMessageConverter.readEntity(reader);
-	}
-	
-	/**
-	 * This test was added for PLFM-1280.
-	 * @throws JSONObjectAdapterException
-	 */
-	@Test (expected=IllegalArgumentException.class)
-	public void testCreateEntityFromAdapterClassNotFound() throws JSONObjectAdapterException{
-		JSONObjectAdapter adapter = new JSONObjectAdapterImpl();
-		adapter.put("concreteType", "org.sagebionetworks.FakeClass");
-		JSONEntityHttpMessageConverter.createEntityFromeAdapter(adapter);
-	}
-
-	/**
-	 * This test was added for PLFM-1280.
-	 * @throws JSONObjectAdapterException
-	 */
-	@Test (expected=JSONObjectAdapterException.class)
-	public void testCreateEntityFromAdapterBadJSON() throws JSONObjectAdapterException{
-		// Test a valid entity type with a field that does not exist on that type.
-		JSONObjectAdapter adapter = new JSONObjectAdapterImpl();
-		adapter.put("entityType", ExampleEntity.class.getName());
-		adapter.put("notAField", "shoudld not exist");
-		JSONEntityHttpMessageConverter.createEntityFromeAdapter(adapter);
-	}
 	
 	@Test
 	public void testPLFM_2079() throws Exception{
@@ -262,113 +169,6 @@ public class JSONEntityHttpMessageConverterTest {
 		}catch(Exception e){
 			throw new RuntimeException(json,e);
 		}
-	}
-	
-	
-	@Test
-	public void testValidateJSONEntityWithValid() throws Exception {
-		// setup
-		JSONObjectAdapter schema = new JSONObjectAdapterImpl();
-		schema.put("description", "Expect this to fail");
-		JSONObjectAdapter adapter = new JSONObjectAdapterImpl();
-		adapter.put("concreteType", "org.sagebionetworks.repo.model.schema.CreateSchemaRequest");
-		adapter.put("schema", schema);
-		String beforeJsonString = adapter.toJSONString();
-		CreateSchemaRequest entity = new CreateSchemaRequest(adapter);
-		// call under test
-		JSONEntityHttpMessageConverter.validateJSONEntity(entity, beforeJsonString);
-	}
-	
-	
-	@Test
-	public void testValidateJSONEntityWithExtraField() throws Exception {
-		// setup no exception to be thrown
-		JSONObjectAdapter schema = new JSONObjectAdapterImpl();
-		schema.put("description", "Expect this to fail");
-		schema.put("notPartOfSpecification", "random");
-		JSONObjectAdapter adapter = new JSONObjectAdapterImpl();
-		adapter.put("concreteType", "org.sagebionetworks.repo.model.schema.CreateSchemaRequest");
-		adapter.put("schema", schema);
-		String beforeJsonString = adapter.toJSONString();
-		CreateSchemaRequest entity = new CreateSchemaRequest(adapter);
-		// call under test
-		String message = assertThrows(IllegalArgumentException.class, () -> {
-			JSONEntityHttpMessageConverter.validateJSONEntity(entity, beforeJsonString);
-		}).getMessage();
-		assertEquals(message, "JSON Element in Entity is Unsupported: notPartOfSpecification");
-	}
-	
-	
-	@Test
-	public void testValidateJSONEntityWithExtraFieldInEmbeddedSchema() throws Exception {
-		// setup no exception to be thrown
-		JSONObjectAdapter items = new JSONObjectAdapterImpl();
-		items.put("notPartOfSpecification", "random");
-		JSONObjectAdapter schema = new JSONObjectAdapterImpl();
-		// items is a JsonSchema
-		schema.put("items", items);
-		schema.put("description", "Expect this to fail");
-		JSONObjectAdapter adapter = new JSONObjectAdapterImpl();
-		adapter.put("concreteType", "org.sagebionetworks.repo.model.schema.CreateSchemaRequest");
-		adapter.put("schema", schema);
-		String beforeJsonString = adapter.toJSONString();
-		CreateSchemaRequest entity = new CreateSchemaRequest(adapter);
-		// call under test
-		String message = assertThrows(IllegalArgumentException.class, () -> {
-			JSONEntityHttpMessageConverter.validateJSONEntity(entity, beforeJsonString);
-		}).getMessage();
-		assertEquals(message, "JSON Element in Entity is Unsupported: notPartOfSpecification");
-	}
-	
-	@Test
-	public void testValidateJSONEntityWithExtraFieldInArray() throws Exception {
-		// setup no exception to be thrown
-		JSONArrayAdapter allOf = new JSONArrayAdapterImpl();
-		JSONObjectAdapter schemaInArray1 = new JSONObjectAdapterImpl();
-		schemaInArray1.put("notPartOfSpecification", "random");
-		JSONObjectAdapter schemaInArray2 = new JSONObjectAdapterImpl();
-		schemaInArray2.put("description", "this is valid though");
-		allOf.put(0, schemaInArray2);
-		allOf.put(1, schemaInArray1);
-		JSONObjectAdapter schema = new JSONObjectAdapterImpl();
-		// "allOf" is an array of JsonSchemas
-		schema.put("allOf", allOf);
-		schema.put("description", "Expect this to fail");
-		JSONObjectAdapter adapter = new JSONObjectAdapterImpl();
-		adapter.put("concreteType", "org.sagebionetworks.repo.model.schema.CreateSchemaRequest");
-		adapter.put("schema", schema);
-		String beforeJsonString = adapter.toJSONString();
-		CreateSchemaRequest entity = new CreateSchemaRequest(adapter);
-		// call under test
-		String message = assertThrows(IllegalArgumentException.class, () -> {
-			JSONEntityHttpMessageConverter.validateJSONEntity(entity, beforeJsonString);
-		}).getMessage();
-		assertEquals(message, "JSON Element in Entity is Unsupported: notPartOfSpecification");
-	}
-	
-	@Test
-	public void testValidateJSONEntityWithExtraFieldInMap() throws Exception {
-		// setup no exception to be thrown
-		JSONObjectAdapter properties = new JSONObjectAdapterImpl();
-		JSONObjectAdapter schema1 = new JSONObjectAdapterImpl();
-		schema1.put("description", "Expect this to fail");
-		JSONObjectAdapter schema2 = new JSONObjectAdapterImpl();
-		schema2.put("notPartOfSpecification", "random");
-		properties.put("schema1", schema1);
-		properties.put("schema2", schema2);
-		JSONObjectAdapter schema = new JSONObjectAdapterImpl();
-		// properties is a map of String to JsonSchema
-		schema.put("properties", properties);
-		JSONObjectAdapter adapter = new JSONObjectAdapterImpl();
-		adapter.put("schema", schema);
-		adapter.put("concreteType", "org.sagebionetworks.repo.model.schema.CreateSchemaRequest");
-		String beforeJsonString = adapter.toJSONString();
-		CreateSchemaRequest entity = new CreateSchemaRequest(adapter);
-		// call under test
-		String message = assertThrows(IllegalArgumentException.class, () -> {
-			JSONEntityHttpMessageConverter.validateJSONEntity(entity, beforeJsonString);
-		}).getMessage();
-		assertEquals(message, "JSON Element in Entity is Unsupported: notPartOfSpecification");
 	}
 	
 	@Test
@@ -409,179 +209,5 @@ public class JSONEntityHttpMessageConverterTest {
 		// call under test
 		CreateSchemaRequest result = (CreateSchemaRequest)nonValidatingConverter.read(CreateSchemaRequest.class, mockInMessage);
 		assertNotNull(result);
-	}
-	
-	@Test
-	public void testValidateJSONEntityWithRequired() throws Exception {
-		// setup
-		JsonSchema schema = new JsonSchema();
-		schema.setDescription("test description");
-		schema.setRequired(Arrays.asList("one", "two"));
-		JSONObjectAdapter schemaAdapter = new JSONObjectAdapterImpl();
-		schema.writeToJSONObject(schemaAdapter);
-		
-		JSONObjectAdapter adapter = new JSONObjectAdapterImpl();
-		adapter.put("concreteType", "org.sagebionetworks.repo.model.schema.CreateSchemaRequest");
-		adapter.put("schema", schemaAdapter);
-		String beforeJsonString = adapter.toJSONString();
-		CreateSchemaRequest entity = new CreateSchemaRequest(adapter);
-		// call under test
-		JSONEntityHttpMessageConverter.validateJSONEntity(entity, beforeJsonString);
-	}
-	
-	@Test
-	public void testValidateJSONEntityRecursiveWithMissingRequiredElement() throws Exception {
-		// this test covers arrays of strings, missing element in array
-		// setup
-		JsonSchema parsedSchema = new JsonSchema();
-		parsedSchema.setDescription("test description");
-		parsedSchema.setRequired(Arrays.asList("one", "two"));
-		JSONObjectAdapter parsedSchemaAdapter = new JSONObjectAdapterImpl();
-		parsedSchema.writeToJSONObject(parsedSchemaAdapter);
-		JSONObjectAdapter parsedAdapter = new JSONObjectAdapterImpl();
-		parsedAdapter.put("concreteType", "org.sagebionetworks.repo.model.schema.CreateSchemaRequest");
-		parsedAdapter.put("schema", parsedSchemaAdapter);
-		
-		JsonSchema originalSchema = new JsonSchema();
-		originalSchema.setDescription("test description");
-		originalSchema.setRequired(Arrays.asList("one", "two", "three"));
-		JSONObjectAdapter originalSchemaAdapter = new JSONObjectAdapterImpl();
-		originalSchema.writeToJSONObject(originalSchemaAdapter);
-		JSONObjectAdapter originalAdapter = new JSONObjectAdapterImpl();
-		originalAdapter.put("concreteType", "org.sagebionetworks.repo.model.schema.CreateSchemaRequest");
-		originalAdapter.put("schema", originalSchemaAdapter);
-
-		String message = assertThrows(IllegalArgumentException.class, () -> { 
-			JSONEntityHttpMessageConverter.validateJSONEntityRecursive(parsedAdapter, originalAdapter);
-		}).getMessage();
-		assertEquals("Missing element in child array of required element on conversion", message);
-	}
-	
-	@Test
-	public void testValidateJSONEntityRecursiveWithIntegerArray() throws Exception {
-		// setup
-		int firstVal = 1;
-		int secondVal = 2;
-		JSONObjectAdapter parsedAdapter = new JSONObjectAdapterImpl();
-		JSONArrayAdapter parsedArray = new JSONArrayAdapterImpl();
-		parsedArray.put(0, firstVal);
-		parsedArray.put(1, secondVal);
-		parsedAdapter.put("arrayKey", parsedArray);
-		
-		JSONObjectAdapter originalAdapter = new JSONObjectAdapterImpl();
-		JSONArrayAdapter originalArray = new JSONArrayAdapterImpl();
-		originalArray.put(0, firstVal);
-		originalArray.put(1, secondVal);
-		originalAdapter.put("arrayKey", originalArray);
-		
-		// call under test
-		JSONEntityHttpMessageConverter.validateJSONEntityRecursive(parsedAdapter, originalAdapter);
-	}
-	
-	@Test
-	public void testValidateJSONEntityRecursiveWithInvalidArrayOfIntegers() throws Exception {
-		// setup
-		int firstVal = 1;
-		int secondVal = 2;
-		JSONObjectAdapter parsedAdapter = new JSONObjectAdapterImpl();
-		JSONArrayAdapter parsedArray = new JSONArrayAdapterImpl();
-		parsedArray.put(0, firstVal);
-		parsedAdapter.put("arrayKey", parsedArray);
-		
-		JSONObjectAdapter originalAdapter = new JSONObjectAdapterImpl();
-		JSONArrayAdapter originalArray = new JSONArrayAdapterImpl();
-		originalArray.put(0, firstVal);
-		originalArray.put(1, secondVal);
-		originalAdapter.put("arrayKey", originalArray);
-		
-		// call under test
-		String message = assertThrows(IllegalArgumentException.class, () -> {
-			JSONEntityHttpMessageConverter.validateJSONEntityRecursive(parsedAdapter, originalAdapter);
-		}).getMessage();
-		
-		assertEquals("Missing element in child array of arrayKey element on conversion", message);
-	}
-	
-	@Test
-	public void testValidateJSONEntityRecursiveWithArrayOfArrays() throws Exception {
-		// set up
-		// The following object is being set up: { "arrayKey": [[], [true, false]] }
-		JSONObjectAdapter parsedAdapter = new JSONObjectAdapterImpl();
-		JSONArrayAdapter parsedArray = new JSONArrayAdapterImpl();
-		JSONArrayAdapter innerParsedArrayOne = new JSONArrayAdapterImpl();
-		JSONArrayAdapter innerParsedArrayTwo = new JSONArrayAdapterImpl();
-		innerParsedArrayTwo.put(0, true);
-		innerParsedArrayTwo.put(1, false);
-		parsedArray.put(0, innerParsedArrayOne);
-		parsedArray.put(1, innerParsedArrayTwo);
-		parsedAdapter.put("arrayKey", parsedArray);
-		
-		JSONObjectAdapter originalAdapter = new JSONObjectAdapterImpl();
-		JSONArrayAdapter originalArray = new JSONArrayAdapterImpl();
-		JSONArrayAdapter innerOriginalArrayOne = new JSONArrayAdapterImpl();
-		JSONArrayAdapter innerOriginalArrayTwo = new JSONArrayAdapterImpl();
-		innerOriginalArrayTwo.put(0, true);
-		innerOriginalArrayTwo.put(1, false);
-		originalArray.put(0, innerOriginalArrayOne);
-		originalArray.put(1, innerOriginalArrayTwo);
-		originalAdapter.put("arrayKey", originalArray);
-		
-		// call under test
-		JSONEntityHttpMessageConverter.validateJSONEntityRecursive(parsedAdapter, originalAdapter);
-	}
-	
-	@Test
-	public void testValidateJSONEntityRecursiveWithMissingArrayInArrayOfArrays() throws Exception {
-		// setup
-		// parsedAdapter: { "arrayKey": [[]] }
-		// originalAdapter: { "arrayKey": [[], [true, false]] }
-		JSONObjectAdapter parsedAdapter = new JSONObjectAdapterImpl();
-		JSONArrayAdapter parsedArray = new JSONArrayAdapterImpl();
-		JSONArrayAdapter innerParsedArrayOne = new JSONArrayAdapterImpl();
-		parsedArray.put(0, innerParsedArrayOne);
-		parsedAdapter.put("arrayKey", parsedArray);
-		
-		JSONObjectAdapter originalAdapter = new JSONObjectAdapterImpl();
-		JSONArrayAdapter originalArray = new JSONArrayAdapterImpl();
-		JSONArrayAdapter innerOriginalArrayOne = new JSONArrayAdapterImpl();
-		JSONArrayAdapter innerOriginalArrayTwo = new JSONArrayAdapterImpl();
-		innerOriginalArrayTwo.put(0, true);
-		innerOriginalArrayTwo.put(1, false);
-		originalArray.put(0, innerOriginalArrayOne);
-		originalArray.put(1, innerOriginalArrayTwo);
-		originalAdapter.put("arrayKey", originalArray);
-		
-		// call under test
-		String message = assertThrows(IllegalArgumentException.class, () -> {
-			JSONEntityHttpMessageConverter.validateJSONEntityRecursive(parsedAdapter, originalAdapter);
-		}).getMessage();
-		
-		assertEquals("Missing element in child array of arrayKey element on conversion", message);
-	}
-	
-	@Test
-	public void testValidateJSONEntityRecursiveWithValidArrayOfJSONObjects() throws Exception {
-		// set up
-		// The following object is being set up: { "arrayKey": [[], [true, false]] }
-		JSONObjectAdapter parsedAdapter = new JSONObjectAdapterImpl();
-		JSONArrayAdapter parsedArray = new JSONArrayAdapterImpl();
-		JSONObjectAdapter parsedObjectOne = new JSONObjectAdapterImpl();
-		JSONObjectAdapter parsedObjectTwo = new JSONObjectAdapterImpl();
-		parsedObjectOne.put("stringKey", "value");
-		parsedArray.put(0, parsedObjectOne);
-		parsedArray.put(1, parsedObjectTwo);
-		parsedAdapter.put("arrayKey", parsedArray);
-		
-		JSONObjectAdapter originalAdapter = new JSONObjectAdapterImpl();
-		JSONArrayAdapter originalArray = new JSONArrayAdapterImpl();
-		JSONObjectAdapter originalObjectOne = new JSONObjectAdapterImpl();
-		JSONObjectAdapter originalObjectTwo = new JSONObjectAdapterImpl();
-		originalObjectOne.put("stringKey", "value");
-		originalArray.put(0, originalObjectOne);
-		originalArray.put(1, originalObjectTwo);
-		originalAdapter.put("arrayKey", originalArray);
-		
-		// call under test
-		JSONEntityHttpMessageConverter.validateJSONEntityRecursive(parsedAdapter, originalAdapter);
 	}
 }
