+++ services/repository/src/main/resources/private/service-spb.xml
@@ -28,6 +28,9 @@
 	<bean id="jwtTypeSerializer"
 		class="org.sagebionetworks.repo.web.controller.JWTTypeSerializerImpl" />
 
+	<bean id="formEncodedTypeSerializer"
+		class="org.sagebionetworks.repo.web.controller.FormEncodedTypeSerializer" />
+
 	<!-- Make sure we can watch for deadlock on all methods of the Generic Entity 
 		Controller -->
 	<bean id="deadlockWatcher"
